Create Elk.sh
While working with this file, I tried to add a chmod +x to it and run it using the "./" command. Due to conflicts with root permissions (despite being logged in as root), I found that the "sh" command (eg. "sh Elk.sh") was the optimal way to run it.

This particular file will appear to stop some of the ways through when it's setting up the connection to the Elk server, and if you ctrl+c out of it, you will not be able to connect to kibana. However, with the network rules in place, navigating to http://[Elk-VM-public-IP]:5601/app/kibana works while the script is actively running.